% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-biopac.R
\name{hasBioread}
\alias{hasBioread}
\title{Is a BIOPAC (bioread) backend available?}
\usage{
hasBioread()
}
\value{
\code{TRUE} when the Python \code{bioread} module is reachable via
\pkg{reticulate}.
}
\description{
Is a BIOPAC (bioread) backend available?
}
\examples{
hasBioread()
}
\seealso{
\code{\link[=readBIOPAC]{readBIOPAC()}}
}
